Split out the Game History Table.
