+++ src/views/User/User.tsx
@@ -27,14 +27,14 @@ import { PlayerIcon } from "PlayerIcon";
 import { GameList } from "GameList";
 import { Player } from "Player";
 import * as preferences from "preferences";
-import { updateDup, getGameResultText, ignore } from "misc";
+import { updateDup, ignore } from "misc";
 import { ModTools } from "./ModTools";
+import { GameHistoryTable } from "./GameHistoryTable";
 import {
     longRankString,
     rankString,
     getUserRating,
     humble_rating,
-    effective_outcome,
     rating_to_rank,
     boundedRankString,
     rank_deviation,
@@ -42,7 +42,7 @@ import {
 import { durationString, daysOnlyDurationString } from "TimeControl";
 import { openModerateUserModal } from "ModerateUser";
 import { PaginatedTable } from "PaginatedTable";
-import { errorAlerter, shouldOpenNewTab } from "misc";
+import { errorAlerter } from "misc";
 import * as player_cache from "player_cache";
 import { getPrivateChat } from "PrivateChat";
 import { PlayerAutocomplete } from "PlayerAutocomplete";
@@ -53,7 +53,7 @@ import { Markdown } from "Markdown";
 import { RatingsChart } from "RatingsChart";
 import { RatingsChartByGame } from "RatingsChartByGame";
 import { associations } from "associations";
-import { browserHistory } from "ogsHistory";
+import { openUrlIfALinkWasNotClicked } from "./common";
 import { Toggle } from "Toggle";
 import { AchievementList } from "Achievements";
 import swal from "sweetalert2";
@@ -68,43 +68,6 @@ interface UserProperties {
 
 const inlineBlock = { display: "inline-flex", alignItems: "center" };
 const marginBottom0 = { marginBottom: "0" };
-
-function getGameResultRichText(game) {
-    let resultText = getGameResultText(game.outcome, game.white_lost, game.black_lost);
-
-    if (game.ranked) {
-        resultText += ", ";
-        resultText += _("ranked");
-    }
-    if (game.annulled) {
-        return (
-            <span>
-                <span style={{ textDecoration: "line-through" }}>{resultText}</span>
-                <span>, {_("annulled")}</span>
-            </span>
-        );
-    }
-
-    return <>{resultText}</>;
-}
-
-function openUrlIfALinkWasNotClicked(ev, url: string) {
-    if (ev.target.nodeName === "A" || ev.target.parentNode.nodeName === "A") {
-        /* if a link was clicked, let the browser handle that. */
-        return;
-    }
-
-    /* Only navigate on left and middle clicks */
-    if (ev.button !== 0 && ev.button !== 1) {
-        return;
-    }
-
-    if (shouldOpenNewTab(ev)) {
-        window.open(url, "_blank");
-    } else {
-        browserHistory.push(url);
-    }
-}
 
 // API: players/%%/full
 interface UserInfo {
@@ -639,14 +602,8 @@ export class User extends React.PureComponent<UserProperties, UserState> {
         });
     };
 
-    maskedRank = (rank: string): string => (preferences.get("hide-ranks") ? "" : rank);
-
     updateTogglePosition = (_height: number, width: number) => {
         this.setState({ rating_chart_type_toggle_left: width + 30 }); // eyeball enough extra left pad
-    };
-
-    onToggleRengoHistorySelect = () => {
-        this.setState({ show_rengo_game_history: !this.state.show_rengo_game_history });
     };
 
     review_history_groomer = (results) => {
@@ -706,133 +663,6 @@ export class User extends React.PureComponent<UserProperties, UserState> {
             }
         };
         setTimeout(doDomWork, 0);
-
-        const game_history_groomer = (results) => {
-            const ret = [];
-            for (let i = 0; i < results.length; ++i) {
-                const r = results[i];
-                const item: any = {
-                    id: r.id,
-                };
-
-                item.width = r.width;
-                item.height = r.height;
-                item.date = r.ended ? new Date(r.ended) : null;
-                item.ranked = r.ranked;
-                item.handicap = r.handicap;
-                item.annulled = r.annulled || false;
-                item.black = r.players.black;
-                item.black_won = !r.black_lost && r.white_lost && !r.annulled;
-                item.black_class = item.black_won
-                    ? item.black.id === this.user_id
-                        ? "library-won"
-                        : "library-lost"
-                    : "";
-                item.white = r.players.white;
-                item.white_won = !r.white_lost && r.black_lost && !r.annulled;
-                item.white_class = item.white_won
-                    ? item.white.id === this.user_id
-                        ? "library-won"
-                        : "library-lost"
-                    : "";
-                item.historical = r.historical_ratings;
-
-                const outcome = effective_outcome(
-                    item.historical.black.ratings.overall.rating,
-                    item.historical.white.ratings.overall.rating,
-                    item.handicap,
-                );
-                if (item.white.id === this.user_id) {
-                    /* played white */ item.played_black = false;
-                    item.opponent = r.historical_ratings.black;
-                    item.player = r.historical_ratings.white;
-                    item.player_won = item.white_won;
-                    if (item.ranked && !preferences.get("hide-ranks")) {
-                        if (item.white_won) {
-                            /* player won */ item.result_class = outcome.white_effective_stronger
-                                ? "library-won-result-vs-weaker"
-                                : "library-won-result-vs-stronger";
-                        } else if (item.black_won) {
-                            /* player lost */ item.result_class = outcome.white_effective_stronger
-                                ? "library-lost-result-vs-weaker"
-                                : "library-lost-result-vs-stronger";
-                        }
-                    } else {
-                        item.result_class = item.white_won
-                            ? "library-won-result-unranked"
-                            : "library-lost-result-unranked"; /* tie catched above */
-                    }
-                } else if (item.black.id === this.user_id) {
-                    /* played black */ item.played_black = true;
-                    item.opponent = r.historical_ratings.white;
-                    item.player = r.historical_ratings.black;
-                    item.player_won = item.black_won;
-                    if (item.ranked && !preferences.get("hide-ranks")) {
-                        if (item.black_won) {
-                            /* player won */ item.result_class = outcome.black_effective_stronger
-                                ? "library-won-result-vs-weaker"
-                                : "library-won-result-vs-stronger";
-                        } else if (item.white_won) {
-                            /* player lost */ item.result_class = outcome.black_effective_stronger
-                                ? "library-lost-result-vs-weaker"
-                                : "library-lost-result-vs-stronger";
-                        }
-                    } else {
-                        item.result_class = item.black_won
-                            ? "library-won-result-unranked"
-                            : "library-lost-result-unranked"; /* tie catched above */
-                    }
-                }
-
-                if (
-                    (r.white_lost && r.black_lost) ||
-                    (!r.white_lost && !r.black_lost) ||
-                    r.annulled
-                ) {
-                    item.result_class = "library-tie-result";
-                }
-
-                if ("time_control_parameters" in r) {
-                    const tcp = JSON.parse(r.time_control_parameters);
-                    if (tcp && "speed" in tcp) {
-                        item.speed = tcp.speed[0].toUpperCase() + tcp.speed.slice(1); // capitalize string
-                    }
-                }
-                if (!item.speed) {
-                    // fallback
-                    if (r.time_per_move >= 3600 || r.time_per_move === 0) {
-                        item.speed = "Correspondence";
-                    } else if (r.time_per_move < 10 && r.time_per_move > 0) {
-                        item.speed = "Blitz";
-                    } else if (r.time_per_move < 3600 && r.time_per_move >= 10) {
-                        item.speed = "Live";
-                    } else {
-                        console.log("time_per_move < 0");
-                    }
-                }
-                if (item.speed === "Correspondence") {
-                    item.speed_icon_class = "speed-icon ogs-turtle";
-                } else if (item.speed === "Blitz") {
-                    item.speed_icon_class = "speed-icon fa fa-bolt";
-                } else if (item.speed === "Live") {
-                    item.speed_icon_class = "speed-icon fa fa-clock-o";
-                } else {
-                    console.log("unsupported speed setting: " + item.speed);
-                }
-
-                item.name = r.name;
-
-                if (item.name && item.name.trim() === "") {
-                    item.name = item.href;
-                }
-
-                item.href = "/game/" + item.id;
-                item.result = getGameResultRichText(r);
-
-                ret.push(item);
-            }
-            return ret;
-        };
 
         let cleaned_website = "";
         if (user && user.website) {
@@ -1225,281 +1055,7 @@ export class User extends React.PureComponent<UserProperties, UserState> {
                         />
 
                         <div className="row">
-                            {/* Game History  */}
-                            <div className="col-sm-12">
-                                <h2>{_("Game History")}</h2>
-                                <Card>
-                                    <div>
-                                        {/* loading-container="game_history.settings().$loading" */}
-                                        <div className="game-options">
-                                            <div className="search">
-                                                <i className="fa fa-search"></i>
-                                                <PlayerAutocomplete
-                                                    onComplete={(player) => {
-                                                        // happily, and importantly, if there isn't a player, then we get null
-                                                        this.setState({
-                                                            games_alt_player_filter: player?.id,
-                                                        });
-                                                    }}
-                                                />
-                                            </div>
-                                            <div className="rengo-selector">
-                                                <span>{_("Rengo")}</span>
-                                                <input
-                                                    type="checkbox"
-                                                    checked={this.state.show_rengo_game_history}
-                                                    onChange={this.onToggleRengoHistorySelect}
-                                                />
-                                            </div>
-                                        </div>
-                                        <PaginatedTable
-                                            className="game-history-table"
-                                            name="game-history"
-                                            method="GET"
-                                            source={`players/${this.user_id}/games/`}
-                                            filter={{
-                                                source: "play",
-                                                ended__isnull: false,
-                                                ...(this.state.games_alt_player_filter !== null && {
-                                                    alt_player: this.state.games_alt_player_filter,
-                                                }),
-                                                rengo: this.state.show_rengo_game_history,
-                                            }}
-                                            orderBy={["-ended"]}
-                                            groom={game_history_groomer}
-                                            onRowClick={(ref, ev) =>
-                                                openUrlIfALinkWasNotClicked(ev, ref.href)
-                                            }
-                                            columns={[
-                                                // normal table layout ...
-                                                ...(!this.state.show_rengo_game_history
-                                                    ? [
-                                                          {
-                                                              header: _("User"),
-                                                              className: (X) =>
-                                                                  "user_info" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <React.Fragment>
-                                                                      {X.played_black ? (
-                                                                          <span>⚫</span>
-                                                                      ) : (
-                                                                          <span>⚪</span>
-                                                                      )}
-                                                                      {this.maskedRank(
-                                                                          `[${rankString(
-                                                                              X.player,
-                                                                          )}]`,
-                                                                      )}
-                                                                  </React.Fragment>
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _(""),
-                                                              className: (X) =>
-                                                                  "winner_marker" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) =>
-                                                                  X.player_won ? (
-                                                                      <i className="fa fa-trophy game-history-winner" />
-                                                                  ) : (
-                                                                      ""
-                                                                  ),
-                                                          },
-                                                          {
-                                                              header: _("Date"),
-                                                              className: (X) =>
-                                                                  "date" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) =>
-                                                                  moment(X.date).format(
-                                                                      "YYYY-MM-DD",
-                                                                  ),
-                                                          },
-                                                          {
-                                                              header: _("Opponent"),
-                                                              className: (X) =>
-                                                                  "player" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <Player
-                                                                      user={X.opponent}
-                                                                      disableCacheUpdate
-                                                                  />
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _(""),
-                                                              className: (X) =>
-                                                                  "speed" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <i
-                                                                      className={X.speed_icon_class}
-                                                                      title={X.speed}
-                                                                  />
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _("Size"),
-                                                              className: (X) =>
-                                                                  "board_size" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) =>
-                                                                  `${X.width}x${X.height}`,
-                                                          },
-                                                          {
-                                                              header: _("Name"),
-                                                              className: (X) =>
-                                                                  "game_name" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <Link to={X.href}>
-                                                                      {X.name ||
-                                                                          interpolate(
-                                                                              "{{black_username}} vs. {{white_username}}",
-                                                                              {
-                                                                                  black_username:
-                                                                                      X.black
-                                                                                          .username,
-                                                                                  white_username:
-                                                                                      X.white
-                                                                                          .username,
-                                                                              },
-                                                                          )}
-                                                                  </Link>
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _("Result"),
-                                                              className: (X) =>
-                                                                  X
-                                                                      ? X.result_class +
-                                                                        (X.annulled
-                                                                            ? " annulled"
-                                                                            : "")
-                                                                      : "",
-                                                              render: (X) => X.result,
-                                                          },
-                                                      ]
-                                                    : []),
-                                                // .. and brute force hiding things that are too hard for a quick implemetation for rengo :o  ...
-                                                ...(this.state.show_rengo_game_history
-                                                    ? [
-                                                          {
-                                                              header: _("-"),
-                                                              className: (X) =>
-                                                                  "user_info" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: () => "",
-                                                          },
-                                                          {
-                                                              header: _(""),
-                                                              className: (X) =>
-                                                                  "winner_marker" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: () => "",
-                                                          },
-                                                          {
-                                                              header: _("Date"),
-                                                              className: (X) =>
-                                                                  "date" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) =>
-                                                                  moment(X.date).format(
-                                                                      "YYYY-MM-DD",
-                                                                  ),
-                                                          },
-                                                          {
-                                                              header: _("-"),
-                                                              className: (X) =>
-                                                                  "player" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: () => "",
-                                                          },
-                                                          {
-                                                              header: _(""),
-                                                              className: (X) =>
-                                                                  "speed" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <i
-                                                                      className={X.speed_icon_class}
-                                                                      title={X.speed}
-                                                                  />
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _("Size"),
-                                                              className: (X) =>
-                                                                  "board_size" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) =>
-                                                                  `${X.width}x${X.height}`,
-                                                          },
-                                                          {
-                                                              header: _("Name"),
-                                                              className: (X) =>
-                                                                  "game_name" +
-                                                                  (X && X.annulled
-                                                                      ? " annulled"
-                                                                      : ""),
-                                                              render: (X) => (
-                                                                  <Link to={X.href}>
-                                                                      {X.name ||
-                                                                          interpolate(
-                                                                              "{{black_username}} vs. {{white_username}}",
-                                                                              {
-                                                                                  black_username:
-                                                                                      X.black
-                                                                                          .username,
-                                                                                  white_username:
-                                                                                      X.white
-                                                                                          .username,
-                                                                              },
-                                                                          )}
-                                                                  </Link>
-                                                              ),
-                                                          },
-                                                          {
-                                                              header: _("Result"),
-                                                              className: () => " annulled",
-                                                              /* a nice background color for this... */ render: () =>
-                                                                  _("fun was had"),
-                                                          },
-                                                      ]
-                                                    : []),
-                                            ]}
-                                        />
-                                    </div>
-                                </Card>
-                            </div>
+                            <GameHistoryTable user_id={this.user_id} />
                         </div>
 
                         <div className="row">
